+++ evalml/pipelines/components/transformers/__init__.py
@@ -31,4 +31,5 @@ from .preprocessing import (
     URLFeaturizer,
     DropRowsTransformer,
     ReplaceNullableTypes,
+    DropNaNRowsTransformer,
 )